requirements.txt
PyAudio==0.2.14
wikipedia==1.4.0
pyttsx3==2.90
SpeechRecognition==3.10.1
requests==2.31.0
python-dotenv==1.0.1